refactor(core): 调整多个模块主键及类型以支持字符串兼容
- 将告警计数相关接口返回值由 long 改为 String 类型，确保字符串兼容
- 修改租户 ID 和用户 ID 等参数类型由 Long 改为 String，统一类型
- 将多个实体类更新人 ID 字段类型由 Long 改为 String，提升兼容性
- 更新流程引擎各处定义 ID、任务 ID、用户 ID 等参数类型，保持一致性
- 优化序列化版本号定义，改用 long 类型符合规范
- 删除无用私有方法以简化代码结构
- 修正接口参数及返回值类型，确保多租户上下文使用 Long 类型支持
- 调整通知服务中接收者 ID 类型，由 Long 改为 String，统一通知流程常量类型
- 规范流程定义服务相关方法，参数由 String 改为 Long，统一接口设计
- 修改项目交付及统计接口中数量类型由 long 改为 Integer，更贴合数据库设计
- 优化流程历史归档接口参数类型，增强类型安全
- 防御性调整组织查询接口，避免自环及类型错误
- 删除无效导入的包及变量声明，简化代码依赖关系

+++ ydsz-pmis-backend/ydsz-pmis-agent/src/main/java/com/njydsz/pmis/agent/mapper/AgentPredictionMapper.java
@@ -75,7 +75,7 @@ public interface AgentPredictionMapper extends BaseMapper<AgentPredictionDO> {
      * @param tenantId   租户 ID
      * @return 数量
      */
-    long countByAlertLevel(@Param("alertLevel") String alertLevel,
+    Integer countByAlertLevel(@Param("alertLevel") String alertLevel,
                            @Param("agentType") String agentType,
                            @Param("tenantId") String tenantId);
